requirements.txt
python_docx==1.1.0
streamlit==1.32.2
